i did a lot of things

+++ Assets/Scripts/Obstacles/Bird.cs
@@ -4,7 +4,7 @@ public class Bird : MonoBehaviour
 {
 
     [SerializeField] private GameObject bird;
-    [SerializeField] private int speed;
+    [SerializeField] private float speed;
     [SerializeField] private float birdMaxDistance;
     void Update()
     {
